# docs/api/openapi/integration-v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "VaniShop Integration API",
        "version": "1",
        "summary": "API cho hệ thống ngoài (ERP, POS, kho, 3PL) đọc đơn/event và ghi xác nhận đơn, tồn kho.",
...
    },
    "servers": [
        {
            "url": "https://{domain}/api/integration/v1",
            "variables": {
                "domain": {
                    "default": "shop.example.vn"
                }
            }
        }
    ],
    "security": [
        {
            "keyId": [],
            "signature": []
        }
    ],
...
    "paths": {
...
        "/inventory/levels": {
            "put": {
                "tags": [
                    "Inventory"
                ],
                "operationId": "syncInventoryLevels",
                "summary": "Tồn vật lý tuyệt đối từ authority (ERP/kho) theo location + SKU",
                "description": "Scope `inventory:write`. Client phải là `stock_authority` của **mọi** location trong request, nếu không cả\nrequest bị từ chối (`403 not_data_owner`). `version` tăng dần theo từng (location, SKU); bản cũ hơn bản đã áp bị\nbỏ qua (`stale_update`). Kết quả trả theo từng dòng.\n",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": false,
                                "required": [
                                    "levels"
                                ],
                                "properties": {
                                    "levels": {
                                        "type": "array",
                                        "minItems": 1,
                                        "maxItems": 500,
                                        "items": {
                                            "type": "object",
                                            "additionalProperties": false,
                                            "required": [
                                                "location_code",
                                                "sku",
                                                "on_hand",
                                                "version"
                                            ],
                                            "properties": {
                                                "location_code": {
                                                    "type": "string",
                                                    "maxLength": 64
                                                },
                                                "sku": {
                                                    "type": "string",
                                                    "maxLength": 64
                                                },
                                                "on_hand": {
                                                    "type": "integer",
                                                    "minimum": 0
                                                },
                                                "version": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kết quả từng dòng.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "../schemas/integration/inventory-levels-result.v1.json"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationFailed"
                    },
                    "429": {
                        "$ref": "#/components/responses/RateLimited"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "keyId": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Vani-Key-Id"
            },
            "signature": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Vani-Signature",
                "description": "t=<unix>,v1=<hex HMAC-SHA256(secret, \"<t>.<METHOD>.<path+query>.<body>\")>"
            }
        },
...
        "responses": {
            "Unauthenticated": {
                "description": "`integration.unauthenticated` — thiếu/sai chữ ký, khoá hết hạn, client tạm dừng.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "../schemas/integration/error.v1.json"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "`integration.insufficient_scope`, `integration.ip_not_allowed`, `not_data_owner`.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "../schemas/integration/error.v1.json"
                        }
                    }
                }
            },
...
            "ValidationFailed": {
                "description": "`validation.failed` — `details` là danh sách {field, messages}.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "../schemas/integration/error.v1.json"
                        }
                    }
                }
            },
            "RateLimited": {
                "description": "`http.429` — vượt giới hạn tần suất của client.",
                "headers": {
                    "Retry-After": {
                        "schema": {
                            "type": "integer"
                        }
                    }
                },
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "../schemas/integration/error.v1.json"
                        }
                    }
                }
            }
        }
    }
}